deploy: 38f9835f43df8889da3198a91f4a174a34bc822d

+++ p5/model.js
@@ -14,7 +14,7 @@ export class Grammar {
 		// convert human friendly to machine suitable representation
 		return sentence.split(" ").map(c => {
 			let op = c[0];
-			let arg = parseInt(c.slice(1)) || undefined;
+			let arg = parseFloat(c.slice(1)) || undefined;
 			return [op, arg];
 		});
 	}
